requirements.txt
flask==3.0.1
gunicorn==21.2.0
Pillow==10.3.0
Requests==2.31.0
torch==2.2.2
transformers==4.36.2
